refactor pipeline for conda; download-tools rules; cluster bug fix; test-data
