+++ config/sample_info.csv.rej
@@ -1,0 +1,46 @@
+diff a/config/sample_info.csv b/config/sample_info.csv	(rejected hunks)
+@@ -1,38 +1,5 @@
+-Sample,PRIDE Accession,Raw file,Raw file URLs,Secondary Sample Accession,Assembly
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669330
+-S27,PXD005780,S27.raw,,ERS1509321,ERZ1669331
+-S51,PXD005780,S51.raw,,ERS1509326,ERZ1669332
+-S61,PXD005780,S61.raw,,ERS1509327,ERZ1669333
+-S23,PXD005780,S23.raw,,ERS1509319,ERZ1669334
+-S25,PXD005780,S25.raw,,ERS1509320,ERZ1669335
+-S25,PXD005780,S25.raw,,ERS1509320,ERZ1669336
+-S13,PXD005780,S13.raw,,ERS1509317,ERZ1669337
+-S36,PXD005780,S36.raw,,ERS1509324,ERZ1669338
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669339
+-S65,PXD005780,S65.raw,,ERS1509328,ERZ1669340
+-S25,PXD005780,S25.raw,,ERS1509320,ERZ1669341
+-S20,PXD005780,S20.raw,,ERS1509318,ERZ1669342
+-S20,PXD005780,S20.raw,,ERS1509318,ERZ1669343
+-S34,PXD005780,S34.raw,,ERS1509323,ERZ1669344
+-S13,PXD005780,S13.raw,,ERS1509317,ERZ1669345
+-S23,PXD005780,S23.raw,,ERS1509319,ERZ1669346
+-S13,PXD005780,S13.raw,,ERS1509317,ERZ1669347
+-S7,PXD005780,S7.raw,,ERS1509316,ERZ1669348
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669349
+-S27,PXD005780,S27.raw,,ERS1509321,ERZ1669350
+-S44,PXD005780,S44.raw,,ERS1509325,ERZ1669351
+-S25,PXD005780,S25.raw,,ERS1509320,ERZ1669352
+-S13,PXD005780,S13.raw,,ERS1509317,ERZ1669353
+-S7,PXD005780,S7.raw,,ERS1509316,ERZ1669354
+-S72,PXD005780,S72.raw,,ERS1509329,ERZ1669355
+-S28,PXD005780,S28.raw,,ERS1509322,ERZ1669356
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669357
+-S20,PXD005780,S20.raw,,ERS1509318,ERZ1669358
+-S20,PXD005780,S20.raw,,ERS1509318,ERZ1669359
+-S23,PXD005780,S23.raw,,ERS1509319,ERZ1669360
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669361
+-S44,PXD005780,S44.raw,,ERS1509325,ERZ1669362
+-S13,PXD005780,S13.raw,,ERS1509317,ERZ1669363
+-S6,PXD005780,S6.raw,,ERS1509315,ERZ1669364
+-S20,PXD005780,S20.raw,,ERS1509318,ERZ1669365
+-S27,PXD005780,S27.raw,,ERS1509321,ERZ1669366
+\ No newline at end of file
++Sample,PRIDE Accession,Raw file,Raw file URLs,Secondary Sample Accession,Assembly
++S6,PXD005780,S6.raw,,ERS1509315,ERZ1669330
++S6,PXD005780,S6.raw,,ERS1509315,ERZ1669339
++S13,PXD005780,S13.raw,,ERS1509317,ERZ1669337
++S13,PXD005780,S13.raw,,ERS1509317,ERZ1669345
