--- workflow/envs/IMP_proteomics.yaml
@@ -1,50 +1,0 @@
-name: IMP_proteomics
-channels:
-  - bioconda
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=0_gnu
-  - alsa-lib=1.1.5=h516909a_1002
-  - fontconfig=2.13.1=h1056068_1002
-  - freetype=2.10.2=he06d7ca_0
-  - gettext=0.19.8.1=hc5be6a0_1002
-  - giflib=5.2.1=h516909a_2
-  - icu=67.1=he1b5a44_0
-  - jpeg=9d=h516909a_0
-  - lcms2=2.9=hbd6801e_2
-  - libffi=3.2.1=he1b5a44_1007
-  - libgcc-ng=9.2.0=h24d8f2e_2
-  - libgomp=9.2.0=h24d8f2e_2
-  - libiconv=1.15=h516909a_1006
-  - libpng=1.6.37=hed695b0_1
-  - libstdcxx-ng=9.2.0=hdf63c60_2
-  - libtiff=4.1.0=hc7e4089_6
-  - libuuid=2.32.1=h14c3975_1000
-  - libwebp-base=1.1.0=h516909a_3
-  - libxcb=1.13=h14c3975_1002
-  - libxml2=2.9.10=h72b56ed_1
-  - lz4-c=1.9.2=he1b5a44_1
-  - mono=5.14.0.177=he1b5a44_1001
-  - openjdk=11.0.1=h600c080_1018
-  - pthread-stubs=0.4=h14c3975_1001
-  - thermorawfileparser=1.2.3=0
-  - xorg-fixesproto=5.0=h14c3975_1002
-  - xorg-inputproto=2.3.2=h14c3975_1002
-  - xorg-kbproto=1.0.7=h14c3975_1002
-  - xorg-libx11=1.6.9=h516909a_0
-  - xorg-libxau=1.0.9=h14c3975_0
-  - xorg-libxdmcp=1.1.3=h516909a_0
-  - xorg-libxext=1.3.4=h516909a_0
-  - xorg-libxfixes=5.0.3=h516909a_1004
-  - xorg-libxi=1.7.10=h516909a_0
-  - xorg-libxrender=0.9.10=h516909a_1002
-  - xorg-libxtst=1.2.3=h516909a_1002
-  - xorg-recordproto=1.14.2=h516909a_1002
-  - xorg-renderproto=0.11.1=h14c3975_1002
-  - xorg-xextproto=7.3.0=h14c3975_1002
-  - xorg-xproto=7.0.31=h14c3975_1007
-  - xz=5.2.5=h516909a_0
-  - zlib=1.2.11=h516909a_1006
-  - zstd=1.4.4=h6597ccf_3
