+++ workflow/envs/pandas_mono.yaml
@@ -1,0 +1,8 @@
+---
+name: pandas
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - mono
+  - pandas=2.0.3
